+++ packages/react-widgets/src/widgets/RangeWidget.js
@@ -1,15 +1,14 @@
-import React, { useCallback, useEffect, useState } from 'react';
-import { useDispatch } from 'react-redux';
+import React, { useCallback, useEffect, useMemo, useState } from 'react';
+import { useDispatch, useSelector } from 'react-redux';
 import { PropTypes } from 'prop-types';
 import { _FilterTypes as FilterTypes } from '@carto/react-core';
 import { WrapperWidgetUI } from '@carto/react-ui';
-import { addFilter } from '@carto/react-redux/';
+import { addFilter, selectSourceById } from '@carto/react-redux/';
 import { getRange } from '../models/RangeModel';
 import useWidgetFetch from '../hooks/useWidgetFetch';
 import WidgetWithAlert from './utils/WidgetWithAlert';
 import { RangeWidgetUI } from '@carto/react-ui';
 import useStats from '../hooks/useStats';
-import { useWidgetFilterValues } from '../hooks/useWidgetFilterValues';
 
 /**
  * Renders a <RangeWidget /> component
@@ -23,7 +22,6 @@ import { useWidgetFilterValues } from '../hooks/useWidgetFilterValues';
  * @param  {boolean} [props.global] - Enable/disable the viewport filtering in the data fetching.
  * @param  {Function} [props.onError] - Function to handle error messages from the widget.
  * @param  {Object} [props.wrapperProps] - Extra props to pass to [WrapperWidgetUI](https://storybook-react.carto.com/?path=/docs/widgets-wrapperwidgetui--default)
- * @param  {Object} [props.droppingFeaturesAlertProps] - Extra props to pass to [NoDataAlert]() when dropping feature
  */
 function RangeWidget({
   id,
@@ -34,10 +32,10 @@ function RangeWidget({
   max: _max,
   global,
   onError,
-  wrapperProps,
-  droppingFeaturesAlertProps
+  wrapperProps
 }) {
   const dispatch = useDispatch();
+  const { filters } = useSelector((state) => selectSourceById(state, dataSource) || {});
   const [[min, max], setMinMax] = useState([_min, _max]);
 
   const hasMinMax = Number.isFinite(min) && Number.isFinite(max);
@@ -49,12 +47,14 @@ function RangeWidget({
     onError
   });
 
-  const selectedValues = useWidgetFilterValues({
-    dataSource,
-    id,
-    column,
-    type: FilterTypes.BETWEEN
-  });
+  // For this widget we cannot use the useWidgetFilterValues hook because this hook
+  // take into account the owner to return the selected values but if two RangeWidget use
+  // the same column they share the selected value because this widget applies
+  // a filter at the time you load it. It doesn't have a "no filter state"
+  const selectedValues = useMemo(
+    () => filters?.[column]?.[FilterTypes.BETWEEN]?.values,
+    [column, filters]
+  );
 
   const {
     data = { min: undefined, max: undefined },
@@ -130,7 +130,7 @@ function RangeWidget({
         dataSource={dataSource}
         warning={warning}
         global={global}
-        droppingFeaturesAlertProps={droppingFeaturesAlertProps}
+        showDroppingFeaturesAlert={false}
       >
         {min !== undefined && max !== undefined && (
           <RangeWidgetUI
@@ -156,8 +156,7 @@ RangeWidget.propTypes = {
   max: PropTypes.number,
   global: PropTypes.bool,
   onError: PropTypes.func,
-  wrapperProps: PropTypes.object,
-  droppingFeaturesAlertProps: PropTypes.object
+  wrapperProps: PropTypes.object
 };
 
 RangeWidget.defaultProps = {