+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Not released
 
+- RangeWidget: new design changes [#459](https://github.com/CartoDB/carto-react/pull/459)
+
 ## 1.4
 
 ### 1.4.0-alpha.2 (2022-07-21)